refactor: move residency counter to op_counter
