+++ spikingjelly/activation_based/op_counter/__init__.py
@@ -1,6 +1,7 @@
 from .base import *
 from .flop import *
 from .memory_access import *
+from .memory_residency import *
 from .ac import *
 from .synop import *
 from .mac import *
